Add a .dockerignore file for the repo root
This commit adds a .dockerignore file at the root of the git repo.
This greatly reduces the size of the build context for people building
from the root of the repo. Of course, it's better to extract the
desired dockerfile to its own directory and build from there, but this
file makes building from the root a little more sane.