Add a Dockerfile for building the NCO from pkgs
The NetCDF Operators are an array of very useful NetCDF
processing tools. Each one individually has a lot of power, and
they can be combined to do pretty much anything that can be done
with a NetCDF file. Working with the operators is usually much
faster than using the C or Python libraries, for example.

A note for Software Engineers:
The NetCDF Operators are very mature, and a single operator is
VERY hard to beat with new code. However, having multiple
NetCDF Operators work on the same files is often less efficient
than a custom-built solution.